Don't lookup user in temporary settings resolvers
The user is not actually required here, we're just interested in the user _ID_ to look up the change. Since a lookup is already made in the auth middleware, just using actor.UID here is just as good. This saves one db call for this query that's run each time the page is visited.