authz/github: allocate less memory for groupsCache disabled path (#24739)
We assign to `seenUsers` and `seenRepos` while syncing users, which can double the memory usage of a sync for no use because the groupsCache disabled code path does not make use of the maps for deduplication (none is necessary). This change removes all assignments not relevant to the groupsCache disabled code path.