Add missing indexes to codeintel tables
There were two really slow queries I found on dotcom; as it turns out, they were missing and index on two columns that were filtered by. This brings down the `HasRepository` query from 580ms to 0.082, and the `IsQueued` query from ~1100ms to 0.141ms.
This also seems to propagate into other codeintel queries as the congestion on these tables is reduced. Overall, we see a 10% CPU usage reduction on dotcom.