Rewrite TemporarySettingsStorage internals
I thought that changing the internals would make it clearer when the
settings are in a loading state. The main idea was to only emit from the
observable if the settings are available, so I rewired some things with
rxjs.

However, given that the local setting backend is the default storage
there isn't actually a loading a state and I'm not sure what we want to
behavior of this to be.

For whatever reason, using rxjs' `delay` function doesn't work in tests.