search: cache results so clicking back button does not re-run search (#24702)
Fixes #24523

Uses React context to cache results based on the options passed to the search function. If the results page is loaded again with the same options, the cached results are shown instead. This significantly improves user experience when navigating results, as they can click the back button and go back to the exact same result list as before.