Fix flaky TestDeleteSnapshots by using fixed point in time (#24833)
As [reported in
Slack](https://sourcegraph.slack.com/archives/C07KZF47K/p1631265491269800),
the test seems to fail when the current time is >24hrs off from the time
stored in the autogold fixture file.

So what I did here is to use a fixed point in time (since `time.Now`
didn't seem relevant to the tests) and save that in the autogold file. I
also updated the other tests to use this point in time.