Surface info on outbound traffic with VPCs
This change is intended to provide an answer to a customer inquiry:

> networking questions!
> I'm assuming out system will need outbound network to
> • our okta installation
> • http://github.comgithub.com> (443 only?)
> • http://sourcegraph.comsourcegraph.com> for telemetry (443 only?)

The intention here is to provide a quick answer for a potential FAQ in the interim. I am hoping CS can later develop an FAQ doc covering network setup questions. We are likely to see more of this as we are now addressing all how-to questions.