Change how WorkspaceResolvers are constructed, simplify interfaces
This is a follow-up to #24886 and hopefully addresses @eseliger's
feedback.

This adds a new interface, `WorkspaceResolverBuilder`, which is a
function that returns a `WorkspaceResolver`.

That makes the flow of the database transaction in the `process` method
itself clearer, since it now allows us to pass such a
`WorkspaceResolverBuilder` into the function - and replace it with a
dummy implementation in the tests.