cmd/gitserver: Move doClone into it's own method (#24885)
This is the first of a series of steps to reduce goroutine creation
for each new clone.

The doClone closure is very involved and while having it defined
withing cloneRepo allows us to not have to worry about argument
passing, it makes understanding the cloneRepo method itself
laborious. As a result, moving it to a method is a good trade off.