Add warning to start.sh that we're switching to sg
This is the first step to make `sg` the default tool to run the
Sourcegraph environment locally.

It prints a message saying that in 2 weeks we'll deprecate the
`start.sh` script and move to `sg` by default.

It also sleeps for 6 seconds to give enough time to read this (and add a
little incentive to switch).