search: Move query state out of root component state (#24844)
This commit moves the "global" query state out of the state of the root component. Because the query state is updated on every keypress it caused excessive re-rendering of the whole application, which, depending on the page, causes massive performance issues.

The query state isn't really "global", rather it's information that is needed by a handful of components rendered in different parts of the page. This commit adds https://zustand.surge.sh/ to make this data available to those components without impacting the overall application (also as a potential solution for other similar use cases).