database: instrument counters on sql requests (#25033)
We were missing metrics around the number of requests we do against
postgresql. This is using the existing sql hook we have for
instrumentating tracing.

In the hook layer we don't know which DB we are accessing, so we have a
single metric for all request and request errors. We use a labeled
metric so it is easy to calculate percentage error.

There is other SQL instrumentation, but that is on the connection /
connection pool.