Remove repositoryQuery results limit workaround
As reported by a customer, and indicated in this issue https://github.com/sourcegraph/sourcegraph/issues/2562, this limit is no longer an issue, therefore this workaround is no longer needed.

Would be good for someone to double-check this is definitely the case, however.

Remove missed paragraphs

Forgot to remove the last couple of paragraphs from this section

Co-authored-by: Christy Haragan <825421+christyharagan@users.noreply.github.com>