API docs: codeintel: begin indexing API docs for search
Depends on #25198 and #25199

Begins writing API docs data to the new `lsif_data_documentation_search_*` tables when
new LSIF bundles are uploaded. This is on by default, but can be disabled via the new
site config feature flag `"apidocs.search.indexing": "disabled"`.

See the architecture docs added in #25197 for more details, and #25199 for details on
the DB schema itself.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>