ssbc: Log final env when resolved (#622)
When determining workspaces, we determine the env, too. The values are preliminary though, as they might depend on what's set in the environment of src-cli or use template variables that aren't known at time of workspace determination. This logs the final set of env vars as used for the step.