cmd/gitserver: Change gauge metrics to counter
We were seeing a flat 0 line for the cloneQueueLength, indicating that
before grafana picks up the metrics, the cloneJobConsumer goroutine
already "processes" the cloneJob and decrements the gauge. Flipping it
to a counter to verify our approach is more reliable.

In any case this is a very short term metric that will be removed in a
few days at best.