search: gob encode zoekt queries once
We currently marshal Q for each replica, which is 49 times for
sourcegraph.com. If you have a large Q this cost can add up to GB of
memory use.

We need to remove most of our large querys we generate, but for now it
is better behaviour to not potentially OOM if we miss something.