search: do not panic / use Zoekt code path if indexed search is disabled
where disabling indexed search via the site config would result in a nil
pointer panic reported in #25269

This was released in 3.32.0, but disabling indexed search is very rare
and generally advised against. We suggest turning indexed search on via
`"search.index.enabled": true,` in your site configuration (or removing
the existing line which disables it) to fix the issue.

Fixes #25269

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>