Delete non-user onboarding test (#25329)
Part of [#inc-2021-09-23-builds-failing-on-main](https://docs.google.com/document/d/1QlosBcYayKqSeMpVuj-TXrFyxmFsiUyZZ_AM5nUz9OU/edit).

We tried to reproduce the test locally, but had difficulty doing so. After looking at the test, we were able to manually verify the conditions (that doing a search + doing a "find references", then reloading the page, causes "Getting started" to no longer appear in the nav bar). That indicates that it's most likely a false positive.

To close the incident and get builds green again, we are temporarily removing this test. Opened #25333 to cover re-enabling the test.

Also adds puppeteer download steps to fix problems in e2e and qa pipelines.