sg: include commit for dev builds
This allows us to warn even if the user's working copy is dirty at time
of installation. We prefix "dev-" to BuildCommit.

Additionally we give advice to users on how to resolve "has changed"
detection failing. This can fail if the user installed sg on a feature
branch whose commits have been GCed.

I noticed this footgun while reading the install.sh script.