fix result streaming on dev environment
Streaming search hasn't been working on the dev environment for a little
while. The issue is the webpack dev server compresses requests with gzip
by default, but gzip buffers the full response before sending it. This
commit disables compression on the dev server so requests to the
`/stream` endpoint are actually streamed back to the client.