Remove `start.sh` and switch to `sg` to run dev environment
2 weeks ago, in #24988, we announced that we'll remove `start.sh` in 2 weeks. Time's up.

This the majority of `start.sh` but leaves a message in it saying that
we switched to `sg` and where to find documentation.

It also removes all the shell scripts that `start.sh` used to boot up
the dev environment that we don't need anymore, either because they're
built into `sg` (such as the watchman configuration, wrapper, helper
scripts) or because they're now in `sg.config.yaml` (where we'll
step-by-step reduce them by building more into `sg`).