sg: check dev-private exists in 'sg start' (#25514)
First attempt at
https://github.com/sourcegraph/sourcegraph/issues/25501.

I think this is fine for now and can avoid new colleagues running into
this problem.

It also got me thinking: I could've defined this as "check" in
`sg.config.yaml`. But then I realised that in order to get a readable
error message out of that I'd have to do some bash scripting. And that
in turn reinforces my thought that the `checks:` should probably be Go
functions. Thoughts?