Return external URLs for Bitbucket Cloud.
For some reason, this wasn't included in the case statement for possible
external URLs, so the frontend was always receiving an empty array.

With this change, I'm able to browse to a Bitbucket Cloud repository
locally and get a generic "View on Bitbucketcloud" option in the UI,
which takes me to the correct location for both the file and at a
specific commit.

Somewhere in the UI there must be another switch statement that needs
a modification to show the icon.