avoid "no results" alert when there are results (#25536)
When streaming results, the aggregator does not hold the list of results
in memory, so the alertObserver was being created with `hasResults ==
false` even though we've streamed results. That caused us to show a "no
results" alert in some cases even though we've sent results.

This commit adds a resultCount field to the aggregator so it can track
how many results it has streamed, and use that for creating the
alertObserver and for generating alerts.

An alternative here would be to add a result count to the Stats struct,
but it felt strange to be updating that when no stats event was sent
down the stream.