remove stdout/stderr redirection from jaeger (#25538)
This commit removes the redirection of stdout/stderr from jaeger when
running in sg. This was causing the bash process that spawned jaeger to
stick around as its parent process, which would catch the kill signal
from a canceled context and not kill its child jaeger process. This
caused jaeger to not be cleaned up on a failed sg start or a panic,
which meant it had to be manually killed if you wanted to restart.