sg: Explicitly define default commandset in config
Using YAML hackery to define the default commandset can lead to subtle
bugs. If someone wants to, for example, overwrite the configuration for
the default set in `sg.config.yaml`, they can't overwrite it under
"default" but need to use the name of the actual command set.

Overwriting what the default set is is also a bit of a pain.

This would make it easier, because setting a new default set would just
require the following in `sg.config.overwrite.yaml`:

    defaultCommandSet: batches