search: support zoekt requesting repo config via ID
It is more efficient to lookup by ID in our database. Additionally we
want to move to being ID based in Zoekt. This commit adds support for
asking for search configuration by repo name or repo id.

A few changes down the road we should be able to avoid building the
intermediate maps of name/id to repo struct. This will require changes
to both Zoekt and Sourcegraph.