database: run schemadoc with postgres 13
This is probably what most people have installed locally. So we use that
instead to avoid the need to run a docker container. Additionally it
should be what is installed on CI, which will speed up the go generate
lint step.

Note: the version check was broken before, it never matched correctly
and would always fall back to docker.