migrations: add dump_root column to lsif_data_documentation_search_* tables
We could rely on the lsif_dumps table for this info, but it's in the main app DB not the codeintel
DB and so would require a separate query or foreign DB wrapper.

We need to know the dump root in order to identify data in the search index that we should purge when
a new upload has been created: all rows which came from a dump for the same upload root and same LSIF
indexer, so we are only keeping search indexes for the latest upload on the default branch of the repo
for each unique (upload_root, language) effectively.

We add `dump_root` as a column to the lsif_data_documentation_search_* table, which will have many
millions of rows. This is duplicative data across rows, similar to the existing `repo_name`, `repo_id`,
`lang`, and often `tags` columns. This is not optimal, but it is OK because the majority of data in this
table (and in trigram indexes) is much larger - duplicating this data is relatively minor and will consist
of less than 5% of the data in this table.

In the near future, once queries against this table have solidified more, I will begin to look at ways of
optimizing this table and storing less information in it so that we can expand the size of the search index
- which is currently restricted to 250 million rows in the `lsif_data_documentation_search_` `_public` and
`_private` tables (this is configurable via a site config option.)

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>