Add original filename to git blame GraphQL response
Previously, the GraphQL endpoint for git blame didn't include the name of the original
file where the blame got added. If the file got renamed, it was not possible to
show the file at the commit where that blame got added. This commit adds
a new `filename: String!` field to the `Hunk` type in the GraphQL
schema.