dev/ci/cluster: deploy to unique BUILDKITE_JOB_ID namespace (#25754)
Using just `BUILDKITE_BUILD_NUMBER` means that namespaces will be used across re-runs of the cluster test job. If you attempt a re-run before the previous cluster has had time to properly spin down, this can cause a re-run to fail.

It seems `BUILDKITE_JOB_ID` is unique across retries so we use that in cluster namespace as well.