convert commit/diff search to job
This creates the type `commit.SearchJob` which captures the minimal set
of validated arguments for commit and diff search and satisfies the
`run.Job` interface. This cuts out all dependency on
`search.TextParameters`, constructing the job directly from a search
query.

This is created in `doResults` instead of `toSearchInputs` because it
depends on having already resolved the list of repos.