commit/diff search: remove formatWithoutRefs (#25705)
This removes formatWithoutRefs, which becomes unused since we always
do commit/diff search with refs. In our current commit/diff code, this
is only used for listing commits (not for searching). If we want to port
the listing code to use the search endpoint, we should implement
skipping refs then, but for now, this change matches the behavior of our
current search paths.