make diffPreview and messagePreview mutually exclusive
A few places in our backend expect message preview and diff preview to
be mutually exclusive in order to detect whether a CommitMatch is
type:commit or type:diff. This is not ideal, but for now, this fixes an
issue with the new commit/diff search path and makes the condition
explicit with a comment.