Rely on MonacoWebpackPlugin for worker config
MonacoWebpackPlugin is actually a _replacement_ for the workers named individually. We can just rely on it for registering all required workers and populating the getWorkerURL function. That makes it easier to align between storybook and the regular web build. Also, this helps make it easier to register additional workers in the future.